test-requirements.txt
# delivery_roulier
odoo_test_helper
vcrpy
